requirements.txt
cryptography==2.1.1
PyQt5==5.9
pyelftools==0.24
